+++ puma/utils/generate.py
@@ -130,13 +130,11 @@ def get_dummy_tagger_aux(
     shuffle: bool = True,
     seed: int = 42,
     label: str = "HadronConeExclTruthLabelID",
-    vtx_label_var="truthVertexIndex",
-    vtx_reco_var="VertexIndex",
 ):
     """
-    Function to generate vertexing aux task output for a tagger, in this case
-    GN2 as well as relevant truth labels. Also includes jet level classifier
-    output and HadronConeExclTruthLabelID.
+    Function to generate aux task output for a tagger, in this case GN2 with
+    vertexing and track origin classification. Also includes jet level
+    classifier output and HadronConeExclTruthLabelID.
 
 
     Parameters
@@ -153,9 +151,7 @@ def get_dummy_tagger_aux(
     label : str, optional
         Name of the label column, by default "HadronConeExclTruthLabelID"
     vtx_label_var : str, optional
-        Name of the truth vertex label, by default "truthVertexIndex"
-    vtx_reco_var : str, optional
-        Name of the reco vertex label, by default "VertexIndex"
+        Name of the truth vertex label, by default "ftagTruthVertexIndex"
 
     Returns
     -------
@@ -170,6 +166,7 @@ def get_dummy_tagger_aux(
 
     rng = np.random.default_rng(seed=seed)
     df_gen["pt"] = rng.exponential(100_000, size=len(df_gen))
+    df_gen["eta"] = rng.normal(0, 2, size=len(df_gen))
     if shuffle:
         df_gen = df_gen.sample(frac=1).reset_index(drop=True)
 
@@ -181,19 +178,25 @@ def get_dummy_tagger_aux(
     vtx_reco = np.fabs(
         np.rint(rng.normal(loc=0, scale=10, size=(len(df_gen), n_tracks))).astype(int)
     )
+    trk_or_labels = np.random.choice(8, size=(len(df_gen), n_tracks)).astype(int)
+    trk_or_reco = np.random.choice(8, size=(len(df_gen), n_tracks)).astype(int)
     aux_dtype = np.dtype(
         [
-            (vtx_label_var, "i4"),
-            (vtx_reco_var, "i4"),
+            ("ftagTruthVertexIndex", "i4"),
+            ("GN2_VertexIndex", "i4"),
+            ("ftagTruthOriginLabel", "i4"),
+            ("GN2_TrackOrigin", "i4"),
         ]
     )
-    vtx_info = np.rec.fromarrays([vtx_labels, vtx_reco], dtype=aux_dtype)
+    aux_info = np.rec.fromarrays(
+        [vtx_labels, vtx_reco, trk_or_labels, trk_or_reco], dtype=aux_dtype
+    )
 
     fname = NamedTemporaryFile(  # pylint: disable=R1732
         mode="w", suffix=".h5", delete=False
     ).name
     file = h5py.File(fname, "w")
     file.create_dataset(name="jets", data=df_gen.to_records())
-    file.create_dataset(name="tracks", data=vtx_info)
+    file.create_dataset(name="tracks", data=aux_info)
 
-    return file
+    return fname, file